Mad Science pack: fix Cheapen's Amplify to actually prevent it from exhausting (and switch it and other effects that change whether the currently played card exhausts to all share code)
