+++ src/main/java/thePackmaster/cards/prismaticpack/PrismaticBarrier.java
@@ -13,6 +13,7 @@ import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import com.megacrit.cardcrawl.powers.PlatedArmorPower;
 import thePackmaster.SpireAnniversary5Mod;
+import thePackmaster.actions.ChangePlayedCardExhaustAction;
 import thePackmaster.cards.AbstractPackmasterCard;
 
 public class PrismaticBarrier extends AbstractPrismaticCard {
@@ -41,20 +42,7 @@ public class PrismaticBarrier extends AbstractPrismaticCard {
         this.addToBot(new GainBlockAction(p, this.block));
         if (this.playedDifferentColorCardCheck()) {
             this.addToBot(new ApplyPowerAction(p, p, new PlatedArmorPower(p, this.magicNumber)));
-            this.addToBot(new AbstractGameAction() {
-                @Override
-                public void update() {
-                    for (AbstractGameAction a : AbstractDungeon.actionManager.actions) {
-                        if (a instanceof UseCardAction) {
-                            if (ReflectionHacks.getPrivate(a, UseCardAction.class, "targetCard") == card) {
-                                ((UseCardAction) a).exhaustCard = true;
-                                break;
-                            }
-                        }
-                    }
-                    this.isDone = true;
-                }
-            });
+            this.addToBot(new ChangePlayedCardExhaustAction(this, true));
         }
     }
